remove unnecesary directory
